Move tests to an appropriate folder
